backend/requirements.txt
# InsightFlow v2 — Backend dependencies (fallback for start.py)
# Canonical spec is pyproject.toml at project root. Keep this in sync.
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
httpx>=0.27.0
python-multipart>=0.0.9
slowapi>=0.1.9
structlog>=24.0.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
aiosqlite>=0.20.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
authlib>=1.3.0
redis[hiredis]>=5.0.0
sqlglot>=25.0.0
sse-starlette>=2.0.0
pydantic-settings>=2.0.0
pydantic[email]>=2.0.0
sentry-sdk[fastapi]>=2.0.0
